Revised release process and changed package management
This change completely revises the release process. Instead of using
craft to build releases with github acitons, the new workflow provides
a Makefile target to build the release locally and push changes out to
master. Not only does this simplify the process (using a bash script),
it also allows for git branches and tags created by the release to be
gpg signed by a human user.

Additionally (to support the release script better), the project was
converted to use poetry for dependency and version management.

CHANGELOG and README were converted to ReStructuredText.